fix timestep
