+++ python/sglang/multimodal_gen/runtime/pipelines/flux.py
@@ -37,26 +37,6 @@ def calculate_shift(
     return mu
 
 
-def compute_empirical_mu(batch: Req, server_args: ServerArgs):
-    num_steps = batch.num_inference_steps
-    image_seq_len = batch.raw_latent_shape[1]
-    a1, b1 = 8.73809524e-05, 1.89833333
-    a2, b2 = 0.00016927, 0.45666666
-
-    if image_seq_len > 4300:
-        mu = a2 * image_seq_len + b2
-        return float(mu)
-
-    m_200 = a2 * image_seq_len + b2
-    m_10 = a1 * image_seq_len + b1
-
-    a = (m_200 - m_10) / 190.0
-    b = m_200 - 200.0 * a
-    mu = a * num_steps + b
-
-    return "mu", float(mu)
-
-
 def prepare_mu(batch: Req, server_args: ServerArgs):
     height = batch.height
     width = batch.width
