+++ python/sglang/multimodal_gen/runtime/pipelines/flux_2.py
@@ -1,8 +1,9 @@
 # Copied and adapted from: https://github.com/hao-ai-lab/FastVideo
+# SPDX-License-Identifier: Apache-2.0
+
 from diffusers.image_processor import VaeImageProcessor
 
-from sglang.multimodal_gen.runtime.pipelines.flux import prepare_mu
-from sglang.multimodal_gen.runtime.pipelines_core import LoRAPipeline
+from sglang.multimodal_gen.runtime.pipelines_core import LoRAPipeline, Req
 from sglang.multimodal_gen.runtime.pipelines_core.composed_pipeline_base import (
     ComposedPipelineBase,
 )
@@ -19,12 +20,27 @@ from sglang.multimodal_gen.runtime.pipelines_core.stages import (
 from sglang.multimodal_gen.runtime.server_args import ServerArgs
 from sglang.multimodal_gen.runtime.utils.logging_utils import init_logger
 
-# SPDX-License-Identifier: Apache-2.0
-
-
-# TODO(will): move PRECISION_TO_TYPE to better place
-
 logger = init_logger(__name__)
+
+
+def compute_empirical_mu(batch: Req, server_args: ServerArgs):
+    num_steps = batch.num_inference_steps
+    image_seq_len = batch.raw_latent_shape[1]
+    a1, b1 = 8.73809524e-05, 1.89833333
+    a2, b2 = 0.00016927, 0.45666666
+
+    if image_seq_len > 4300:
+        mu = a2 * image_seq_len + b2
+        return "mu", float(mu)
+
+    m_200 = a2 * image_seq_len + b2
+    m_10 = a1 * image_seq_len + b1
+
+    a = (m_200 - m_10) / 190.0
+    b = m_200 - 200.0 * a
+    mu = a * num_steps + b
+
+    return "mu", float(mu)
 
 
 class Flux2Pipeline(LoRAPipeline, ComposedPipelineBase):
@@ -62,7 +78,7 @@ class Flux2Pipeline(LoRAPipeline, ComposedPipelineBase):
             stage=ImageVAEEncodingStage(
                 vae_image_processor=VaeImageProcessor(
                     vae_scale_factor=server_args.pipeline_config.vae_config.arch_config.vae_scale_factor
-                    * 2
+                                     * 2
                 ),
                 vae=self.get_module("vae"),
             ),
@@ -71,18 +87,18 @@ class Flux2Pipeline(LoRAPipeline, ComposedPipelineBase):
         self.add_stage(stage_name="conditioning_stage", stage=ConditioningStage())
 
         self.add_stage(
-            stage_name="timestep_preparation_stage",
-            stage=TimestepPreparationStage(
-                scheduler=self.get_module("scheduler"),
-                prepare_extra_set_timesteps_kwargs=[prepare_mu],
-            ),
-        )
-
-        self.add_stage(
             stage_name="latent_preparation_stage",
             stage=LatentPreparationStage(
                 scheduler=self.get_module("scheduler"),
                 transformer=self.get_module("transformer"),
+            ),
+        )
+
+        self.add_stage(
+            stage_name="timestep_preparation_stage",
+            stage=TimestepPreparationStage(
+                scheduler=self.get_module("scheduler"),
+                prepare_extra_set_timesteps_kwargs=[compute_empirical_mu],
             ),
         )
